+++ packages/blocks/src/_common/copilot/model/message-schema.ts
@@ -49,6 +49,7 @@ export type ApiData<T> =
     }
   | {
       status: 'stop';
+      data?: T;
     }
   | {
       status: 'success';
