+++ packages/blocks/src/_common/copilot/model/message-type/text/components/generating.ts
@@ -1,0 +1,66 @@
+import { WithDisposable } from '@blocksuite/block-std';
+import { css, html, LitElement } from 'lit';
+import { customElement, property } from 'lit/decorators.js';
+
+import {
+  AIStarIconWithAnimation,
+  AIStopIcon,
+} from '../../../../../icons/ai.js';
+import type { AssistantHistoryItem } from '../../../chat-history.js';
+
+@customElement('ai-text-generating')
+export class AITextGenerating extends WithDisposable(LitElement) {
+  static override styles = css`
+    .generating-tip {
+      display: flex;
+      width: 100%;
+      height: 22px;
+      align-items: center;
+      gap: 8px;
+
+      color: var(--light-brandColor, #1e96eb);
+
+      .text {
+        display: flex;
+        align-items: flex-start;
+        gap: 10px;
+        flex: 1 0 0;
+
+        /* light/smMedium */
+        font-family: Inter;
+        font-size: 14px;
+        font-style: normal;
+        font-weight: 500;
+        line-height: 22px; /* 157.143% */
+      }
+
+      .left,
+      .right {
+        display: flex;
+        width: 20px;
+        height: 20px;
+        justify-content: center;
+        align-items: center;
+      }
+    }
+  `;
+
+  @property({ attribute: false })
+  item!: AssistantHistoryItem;
+
+  override render() {
+    return html`
+      <div class="generating-tip">
+        <div class="left">${AIStarIconWithAnimation}</div>
+        <div class="text">AI is generating...</div>
+        <div @click=${this.item.stop} class="right">${AIStopIcon}</div>
+      </div>
+    `;
+  }
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'ai-text-generating': AITextGenerating;
+  }
+}
