+++ packages/blocks/src/_common/copilot/model/message-type/text/components/index.ts
@@ -1,0 +1,5 @@
+export * from './actions.js';
+export * from './answer.js';
+export * from './divider.js';
+export * from './error.js';
+export * from './generating.js';
