+++ packages/blocks/src/_common/icons/ai.ts
@@ -1,3 +1,5 @@
+import '@dotlottie/player-component';
+
 import { html } from 'lit';
 
 export const AIStarIcon = html`<svg
@@ -23,6 +25,15 @@ export const AIStarIcon = html`<svg
     </clipPath>
   </defs>
 </svg>`;
+
+export const AIStarIconWithAnimation = html`<dotlottie-player
+  src='{"v":"5.12.1","fr":120,"ip":0,"op":239,"w":240,"h":240,"nm":"ai","ddd":0,"assets":[],"layers":[{"ddd":0,"ind":1,"ty":4,"nm":"big 2","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":38,"s":[80]},{"t":76,"s":[100]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[107.859,132.143,0],"ix":2,"l":2},"a":{"a":0,"k":[8.572,11.429,0],"ix":1,"l":2},"s":{"a":1,"k":[{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":38,"s":[0,0,100]},{"t":76,"s":[1005,1005,100]}],"ix":6,"l":2}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0.261,2.307],[0.363,0],[0.041,-0.361],[1.131,-1.132],[2.346,-0.271],[-0.001,-0.363],[-0.361,-0.041],[-1.158,-1.137],[-0.207,-2.329],[-0.37,0],[-0.032,0.369],[-1.16,1.16],[-2.291,0.198],[-0.001,0.371],[0.368,0.033],[1.133,1.154]],"o":[[-0.041,-0.361],[-0.363,0],[-0.271,2.345],[-1.132,1.131],[-0.36,0.042],[0,0.363],[2.307,0.262],[1.154,1.133],[0.033,0.369],[0.37,0],[0.198,-2.29],[1.16,-1.159],[0.369,-0.032],[0,-0.37],[-2.329,-0.207],[-1.138,-1.158]],"v":[[0.71,-7.937],[0.001,-8.571],[-0.709,-7.938],[-2.812,-2.812],[-7.939,-0.709],[-8.571,0.001],[-7.937,0.71],[-2.788,2.81],[-0.711,7.921],[0.001,8.572],[0.712,7.919],[2.785,2.785],[7.919,0.712],[8.572,0.001],[7.921,-0.711],[2.81,-2.788]],"c":true},"ix":2},"nm":"路径 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.117647058824,0.588235294118,0.921568627451,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"填充 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[8.572,11.429],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"变换"}],"nm":"组 1","np":4,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":8400,"st":0,"ct":1,"bm":0},{"ddd":0,"ind":2,"ty":4,"nm":"big","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":0,"s":[100]},{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":15,"s":[80]},{"t":38,"s":[0]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[107.859,132.143,0],"ix":2,"l":2},"a":{"a":0,"k":[8.572,11.429,0],"ix":1,"l":2},"s":{"a":1,"k":[{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":0,"s":[1005,1005,100]},{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":15,"s":[1035,1035,100]},{"t":38,"s":[0,0,100]}],"ix":6,"l":2}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0.261,2.307],[0.363,0],[0.041,-0.361],[1.131,-1.132],[2.346,-0.271],[-0.001,-0.363],[-0.361,-0.041],[-1.158,-1.137],[-0.207,-2.329],[-0.37,0],[-0.032,0.369],[-1.16,1.16],[-2.291,0.198],[-0.001,0.371],[0.368,0.033],[1.133,1.154]],"o":[[-0.041,-0.361],[-0.363,0],[-0.271,2.345],[-1.132,1.131],[-0.36,0.042],[0,0.363],[2.307,0.262],[1.154,1.133],[0.033,0.369],[0.37,0],[0.198,-2.29],[1.16,-1.159],[0.369,-0.032],[0,-0.37],[-2.329,-0.207],[-1.138,-1.158]],"v":[[0.71,-7.937],[0.001,-8.571],[-0.709,-7.938],[-2.812,-2.812],[-7.939,-0.709],[-8.571,0.001],[-7.937,0.71],[-2.788,2.81],[-0.711,7.921],[0.001,8.572],[0.712,7.919],[2.785,2.785],[7.919,0.712],[8.572,0.001],[7.921,-0.711],[2.81,-2.788]],"c":true},"ix":2},"nm":"路径 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.117647058824,0.588235294118,0.921568627451,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"填充 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[8.572,11.429],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"变换"}],"nm":"组 1","np":4,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":8400,"st":0,"ct":1,"bm":0},{"ddd":0,"ind":3,"ty":4,"nm":"small 2","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":48.074,"s":[80]},{"t":85.07421875,"s":[100]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[183,56.999,0],"ix":2,"l":2},"a":{"a":0,"k":[16.667,3.333,0],"ix":1,"l":2},"s":{"a":1,"k":[{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":48,"s":[0,0,100]},{"t":85.07421875,"s":[1005,1005,100]}],"ix":6,"l":2}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0.102,0.897],[0.141,0.001],[0.016,-0.14],[0.44,-0.44],[0.912,-0.105],[0,-0.141],[-0.14,-0.016],[-0.451,-0.442],[-0.08,-0.906],[-0.144,0],[-0.013,0.144],[-0.451,0.451],[-0.89,0.077],[0,0.144],[0.143,0.013],[0.441,0.448]],"o":[[-0.016,-0.14],[-0.141,0],[-0.105,0.913],[-0.44,0.44],[-0.141,0.016],[0,0.141],[0.897,0.101],[0.448,0.441],[0.013,0.143],[0.144,0],[0.077,-0.891],[0.451,-0.451],[0.144,-0.013],[0,-0.144],[-0.906,-0.08],[-0.442,-0.451]],"v":[[0.276,-3.087],[0,-3.334],[-0.276,-3.088],[-1.094,-1.094],[-3.087,-0.276],[-3.333,0],[-3.087,0.276],[-1.084,1.092],[-0.277,3.08],[0,3.333],[0.277,3.079],[1.083,1.083],[3.079,0.277],[3.333,0],[3.08,-0.277],[1.092,-1.084]],"c":true},"ix":2},"nm":"路径 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.117647058824,0.588235294118,0.921568627451,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"填充 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[16.667,3.333],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"变换"}],"nm":"组 1","np":4,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":8400,"st":0,"ct":1,"bm":0},{"ddd":0,"ind":4,"ty":4,"nm":"small","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":10.074,"s":[100]},{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":25.074,"s":[80]},{"t":48.07421875,"s":[0]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[183,56.999,0],"ix":2,"l":2},"a":{"a":0,"k":[16.667,3.333,0],"ix":1,"l":2},"s":{"a":1,"k":[{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":10.074,"s":[1005,1005,100]},{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":25.074,"s":[1235,1235,100]},{"t":48,"s":[0,0,100]}],"ix":6,"l":2}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0.102,0.897],[0.141,0.001],[0.016,-0.14],[0.44,-0.44],[0.912,-0.105],[0,-0.141],[-0.14,-0.016],[-0.451,-0.442],[-0.08,-0.906],[-0.144,0],[-0.013,0.144],[-0.451,0.451],[-0.89,0.077],[0,0.144],[0.143,0.013],[0.441,0.448]],"o":[[-0.016,-0.14],[-0.141,0],[-0.105,0.913],[-0.44,0.44],[-0.141,0.016],[0,0.141],[0.897,0.101],[0.448,0.441],[0.013,0.143],[0.144,0],[0.077,-0.891],[0.451,-0.451],[0.144,-0.013],[0,-0.144],[-0.906,-0.08],[-0.442,-0.451]],"v":[[0.276,-3.087],[0,-3.334],[-0.276,-3.088],[-1.094,-1.094],[-3.087,-0.276],[-3.333,0],[-3.087,0.276],[-1.084,1.092],[-0.277,3.08],[0,3.333],[0.277,3.079],[1.083,1.083],[3.079,0.277],[3.333,0],[3.08,-0.277],[1.092,-1.084]],"c":true},"ix":2},"nm":"路径 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.117647058824,0.588235294118,0.921568627451,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"填充 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[16.667,3.333],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"变换"}],"nm":"组 1","np":4,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":8400,"st":0,"ct":1,"bm":0}],"markers":[],"props":{}}'
+  background="transparent"
+  speed="1"
+  style="width: 300px; height: 300px;"
+  loop
+  autoplay
+></dotlottie-player>`;
 
 export const AIStopIcon = html`<svg
   width="20"
