+++ packages/blocks/src/root-block/widgets/ai-action-panel/components/index.ts
@@ -1,4 +1,1 @@
-// eslint-disable-next-line simple-import-sort/exports
-export * from './ai-action-panel-base.js';
-
 export * from './ai-action-panel-input.js';
